pyproject.toml
[tool.poetry]
name = "csye7230"
version = "0.1.0"
description = "search2.0"
authors = ["Frieda (Jingying) Huang <jingyingfhuang@gmail.com>"]
readme = "README.md"
packages = [{include = "searchagent"}]


[tool.poetry.dependencies]
python = ">=3.12,<3.13"
pre-commit = "^3.8.0"
ruff = "^0.6.5"
ragas = "^0.1.18"
loguru = "^0.7.2"
pyyaml = "^6.0.2"
langchain = "0.2.11"
langchain-ollama = "0.1.3"
langchain-huggingface = "0.0.3"
python-dotenv = "^1.0.1"
datasets = "^3.0.0"
tqdm = "^4.66.5"
pandas = "^2.2.3"
ipykernel = "^6.29.5"
ragatouille = "^0.0.8.post4"
einops = "^0.8.0"
pgvector = "^0.3.3"
python-magic = "^0.4.27"
pypdf = "^5.0.0"
python-docx = "^1.1.2"
colpali-engine = "^0.3.3"
torch = "^2.4.1"
pdf2image = "^1.17.0"
llama-stack = "0.0.55"
sqlalchemy = "^2.0.35"
alembic = "^1.13.3"
psycopg = {extras = ["binary", "pool"], version = "^3.2.3"}
bcrypt = "^4.2.0"
ml-dtypes = "^0.5.0"
llama-stack-client = "^0.0.55"
watchdog = "^5.0.3"
fastapi = {extras = ["standard"], version = "^0.115.4"}
torchmetrics = "^1.5.1"
pydantic = "^2.9.2"
pydantic-settings = "^2.6.1"
anthropic = "^0.39.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
